Move http_sidecar under http.server
